Done!
As far as I know, this is pretty much done.
-Took away the ability to set a starting price in data.txt. It steps
through 10 iterations to begin with, so it doesn't really affect
anything anyway. It just sets the starting price to the average of the
high and low.
-You can now see the price in multiples of 2-5.
-Made it look prettier!

Like I said, unless there's anything else that really needs to be added,
this is complete. Feel free to tweak the visuals if you think you can
make it look better.